docqa_ocr_albert/requirements.txt
flask==2.3.2
pytesseract==0.3.10
pdfplumber==0.9.0
python-docx==0.8.11
transformers==4.31.0
torch==2.0.1
pdf2image==1.16.3
opencv-python==4.8.0.76
numpy==1.24.3
Pillow==10.0.0